services/test-svc/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
google-genai>=1.3.0
pydantic==2.7.0
